+++ forms/src/components/defaultValidationErrorsContainer/defaultValidationErrorsContainer.component.ts
@@ -10,7 +10,8 @@ import {ValidationErrorsComponent, ValidationErrorsContainerComponent, Validatio
 {
     selector: 'default-validation-errors-container',
     template: '',
-    styleUrls: ['defaultValidationErrorsContainer.component.css'],
+    styleUrl: 'defaultValidationErrorsContainer.component.css',
+    standalone: true,
     changeDetection: ChangeDetectionStrategy.OnPush
 })
 export class DefaultValidationErrorsContainerComponent implements ValidationErrorsContainerComponent