+++ forms/src/directives/groupHasErrorContainer/groupHasErrorContainer.directive.ts
@@ -9,6 +9,7 @@ import {GroupHasErrorDirective} from '../groupHasError/groupHasError.directive';
 @Directive(
 {
     selector: '[groupHasErrorContainer]',
+    standalone: true,
     providers:
     [
         <ExistingProvider>