+++ forms/src/directives/validationErrorsContainer/validationErrorsContainer.directive.ts
@@ -7,7 +7,8 @@ import {ValidationErrorsContainerView} from '../../misc/validationErrorsContaine
  */
 @Directive(
 {
-    selector: '[validationErrorsContainer]'
+    selector: '[validationErrorsContainer]',
+    standalone: true,
 })
 export class ValidationErrorsContainerDirective
 {