+++ forms/src/modules/hasError.module.ts
@@ -1,10 +1,8 @@
 import {NgModule} from '@angular/core';
-import {CommonModule} from '@angular/common';
 
 import {HasErrorDirective} from '../directives/hasError/hasError.directive';
 import {GroupHasErrorDirective} from '../directives/groupHasError/groupHasError.directive';
 import {GroupHasErrorContainerDirective, ValidationErrorsContainerDirective} from '../directives';
-import {DefaultValidationErrorsComponent, DefaultValidationErrorsContainerComponent, ReservedSpaceValidationErrorsContainerComponent} from '../components';
 
 /**
  * Module for input validation directives
@@ -13,17 +11,10 @@ import {DefaultValidationErrorsComponent, DefaultValidationErrorsContainerCompon
 {
     imports:
     [
-        CommonModule,
         HasErrorDirective,
         GroupHasErrorDirective,
-    ],
-    declarations: 
-    [
         GroupHasErrorContainerDirective,
         ValidationErrorsContainerDirective,
-        DefaultValidationErrorsComponent,
-        DefaultValidationErrorsContainerComponent,
-        ReservedSpaceValidationErrorsContainerComponent,
     ],
     exports: 
     [